#301 pulll out browser launching logic
