+++ lib/taiko.js
@@ -1,10 +1,7 @@
 const { doActionAwaitingNavigation } = require('./doActionAwaitingNavigation');
-
 const cri = require('chrome-remote-interface');
-const childProcess = require('child_process');
 const isReachable = require('is-reachable');
 const {
-  helper,
   wait,
   isString,
   isStrictObject,
@@ -54,7 +51,6 @@ const {
   defaultConfig,
   setNavigationOptions,
   setClickOptions,
-  setBrowserOptions,
 } = require('./config');
 const fs = require('fs-extra');
 const path = require('path');
@@ -62,10 +58,13 @@ const { eventHandler } = require('./eventBus');
 const overlayHandler = require('./handlers/overlayHandler');
 const { highlightElement } = require('./elements/elementHelper');
 const numRetries = defaultConfig.criConnectionRetries;
+const {
+  launchChrome,
+  closeChrome,
+  errorMessageForChromeProcessCrash,
+} = require('./browserLauncher');
 
-let chromeProcess,
-  temporaryUserDataDir,
-  page,
+let page,
   network,
   input,
   _client,
@@ -115,7 +114,11 @@ const createProxyForCDPDomain = (cdpClient, cdpDomainName) => {
         return async (...args) => {
           return await new Promise((resolve, reject) => {
             eventHandler.removeAllListeners('browserCrashed');
-            eventHandler.on('browserCrashed', reject);
+            eventHandler.on('browserCrashed', () => {
+              _client.removeAllListeners();
+              clientProxy = null;
+              reject();
+            });
             let res = domainApi.apply(null, args);
             if (res instanceof Promise) {
               res.then(resolve).catch(reject);
@@ -123,11 +126,8 @@ const createProxyForCDPDomain = (cdpClient, cdpDomainName) => {
               resolve(res);
             }
           }).catch((e) => {
-            if (
-              e.message.match(/WebSocket is not open: readyState 3/) &&
-              hasChromeProcessCrashed()
-            ) {
-              throw new Error(errorMessageForChromeprocessCrash());
+            if (e.message.match(/WebSocket is not open: readyState 3/)) {
+              errorMessageForChromeProcessCrash();
             }
             throw e;
           });
@@ -210,10 +210,6 @@ const connect_to_cri = async (target, options = {}) => {
   return initCRI(tgt, numRetries, options);
 };
 
-function hasChromeProcessCrashed() {
-  return chromeProcess && chromeProcess.killed && chromeProcess.exitCode !== 0;
-}
-
 async function reconnect() {
   const response = await isReachable(`${currentHost}:${currentPort}`);
   if (response) {
@@ -251,59 +247,6 @@ eventHandler.addListener('targetCreated', async (newTarget) => {
     });
   }
 });
-
-const browserExitEventHandler = () => {
-  chromeProcess.killed = true;
-  _client.removeAllListeners();
-  clientProxy = null;
-  eventHandler.emit('browserCrashed', new Error(errorMessageForChromeprocessCrash()));
-};
-
-async function setBrowserArgs(options) {
-  let args = [
-    `--remote-debugging-port=${options.port}`,
-    '--disable-features=site-per-process,TranslateUI',
-    '--enable-features=NetworkService,NetworkServiceInProcess',
-    '--disable-renderer-backgrounding',
-    '--disable-backgrounding-occluded-windows',
-    '--disable-background-timer-throttling',
-    '--disable-background-networking',
-    '--disable-breakpad',
-    '--disable-default-apps',
-    '--disable-hang-monitor',
-    '--disable-prompt-on-repost',
-    '--disable-sync',
-    '--force-color-profile=srgb',
-    '--safebrowsing-disable-auto-update',
-    '--password-store=basic',
-    '--use-mock-keychain',
-    '--enable-automation',
-    '--disable-notifications',
-    'about:blank',
-  ];
-  args = options.args ? args.concat(options.args) : args;
-  args = process.env.TAIKO_BROWSER_ARGS
-    ? args.concat(
-        process.env.TAIKO_BROWSER_ARGS.split(/\s*,?\s*--/)
-          .filter((arg) => arg !== '')
-          .map((arg) => `--${arg}`),
-      )
-    : args;
-  if (!args.some((arg) => arg.startsWith('--user-data-dir'))) {
-    const os = require('os');
-    const CHROME_PROFILE_PATH = path.join(os.tmpdir(), 'taiko_dev_profile-');
-    const mkdtempAsync = helper.promisify(fs.mkdtemp);
-    temporaryUserDataDir = await mkdtempAsync(CHROME_PROFILE_PATH);
-    args.push(`--user-data-dir=${temporaryUserDataDir}`);
-  }
-  if (options.headless) {
-    args.push('--headless');
-    if (!args.some((arg) => arg.startsWith('--window-size'))) {
-      args.push('--window-size=1440,900');
-    }
-  }
-  return args;
-}
 
 /**
  * Launches a browser with a tab. The browser will be closed when the parent node.js process is closed.<br>
@@ -343,32 +286,11 @@ module.exports.openBrowser = async (
     );
   }
 
-  if (chromeProcess && !chromeProcess.killed) {
-    throw new Error('OpenBrowser cannot be called again as there is a chromium instance open.');
-  }
-
   if (options.host && options.port) {
     currentHost = options.host;
     currentPort = options.port;
   } else {
-    const BrowserFetcher = require('./browserFetcher');
-    const browserFetcher = new BrowserFetcher();
-    const chromeExecutable = browserFetcher.getExecutablePath();
-    options = setBrowserOptions(options);
-    let args = await setBrowserArgs(options);
-    chromeProcess = await childProcess.spawn(chromeExecutable, args);
-    if (options.dumpio) {
-      chromeProcess.stderr.pipe(process.stderr);
-      chromeProcess.stdout.pipe(process.stdout);
-    }
-    chromeProcess.once('exit', browserExitEventHandler);
-    const endpoint = await browserFetcher.waitForWSEndpoint(
-      chromeProcess,
-      defaultConfig.navigationTimeout,
-    );
-    currentHost = endpoint.host;
-    currentPort = endpoint.port;
-    browserDebugUrl = endpoint.browser;
+    ({ currentHost, currentPort, browserDebugUrl } = await launchChrome(options));
   }
   await connect_to_cri();
   var description = device ? `Browser opened with viewport ${device}` : 'Browser opened';
@@ -419,39 +341,8 @@ const _closeBrowser = async () => {
     await _client.close();
     clientProxy = null;
   }
-  const waitForChromeToClose = new Promise((fulfill) => {
-    chromeProcess.removeAllListeners();
-    chromeProcess.once('exit', () => {
-      fulfill();
-    });
-    if (chromeProcess.killed) {
-      fulfill();
-    }
-    timeout = setTimeout(() => {
-      fulfill();
-      chromeProcess.removeAllListeners();
-      chromeProcess.kill('SIGKILL');
-    }, defaultConfig.retryTimeout);
-  });
-  chromeProcess.kill('SIGTERM');
-  await waitForChromeToClose;
-  clearTimeout(timeout);
-  if (temporaryUserDataDir) {
-    try {
-      fs.removeSync(temporaryUserDataDir);
-    } catch (e) {}
-  }
+  await closeChrome();
 };
-
-function errorMessageForChromeprocessCrash() {
-  let message;
-  if (chromeProcess.exitCode === null) {
-    message = `Chrome process with pid ${chromeProcess.pid} exited with signal ${chromeProcess.signalCode}.`;
-  } else {
-    message = `Chrome process with pid ${chromeProcess.pid} exited with status code ${chromeProcess.exitCode}.`;
-  }
-  return message;
-}
 
 function getEventProxy(target) {
   let unsupportedClientMethods = [
@@ -3384,11 +3275,7 @@ const validate = () => {
     throw new Error('Browser or page not initialized. Call `openBrowser()` before using this API');
   }
   if (_client._ws.readyState > 1) {
-    if (chromeProcess && chromeProcess.killed) {
-      throw new Error(
-        'The Browser instance was closed either via `closeBrowser()` call, or it crashed for reasons unknown to Taiko. You can try launching a fresh instance using `openBrowser()` or inspect the logs for details of the possible crash.',
-      );
-    }
+    errorMessageForChromeProcessCrash();
     throw new Error(
       "Connection to browser lost. This probably isn't a problem with Taiko, inspect logs for possible causes.",
     );
